# funcionarios-api/swagger.yaml
swagger: '2.0'
info:
  description: Api de controle de funcionários
  version: 1.0.0
  title: Funcionários API
  contact:
    name: Charles Luxinger
    url: 'https://www.linkedin.com/in/charlesluxinger/'
    email: charlesluxinger@gmail.com
host: 'localhost:8085'
basePath: /funcionarios-api
tags:
  - name: funcionario-rest-controller
    description: Funcionario Rest Controller
paths:
  /funcionario:
    post:
      tags:
        - funcionario-rest-controller
      summary: cadastra
      operationId: cadastraUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: funcionario
          description: funcionario
          required: true
          schema:
            $ref: '#/definitions/Funcionario'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Funcionario'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /funcionario/todos:
    get:
      tags:
        - funcionario-rest-controller
      summary: todos
      operationId: todosUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Funcionario'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/funcionario/{id}':
    put:
      tags:
        - funcionario-rest-controller
      summary: altera
      operationId: alteraUsingPUT
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
        - in: body
          name: funcionario
          description: funcionario
          required: true
          schema:
            $ref: '#/definitions/Funcionario'
      responses:
        '200':
          description: OK
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - funcionario-rest-controller
      summary: remove
      operationId: removeUsingDELETE
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
definitions:
  Funcionario:
    type: object
    properties:
      dataContratacao:
        type: string
        format: date-time
      dataNascimento:
        type: string
        format: date-time
      id:
        type: integer
        format: int64
      nome:
        type: string
      salario:
        type: number
        format: double